requirements/prod.txt
# Everything that needed in production
Flask==0.11.1
Flask-WTF==0.13.1
Flask-Script==2.0.5
Jinja2==2.8
MarkupSafe==0.23
WTForms==2.1
Werkzeug==0.11.11
gunicorn==19.6.0
itsdangerous==0.24
flask-mongoengine==0.8
mongoengine==0.10.6
pymongo==4.5.0
six==1.10.0
blinker==1.4
PyJWT==2.8.0